move layouts to pages
