--- src/components/appShell/index.ts
@@ -1,1 +1,0 @@
-export * from './AppShell';
