+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './appShell';
 export * from './atoms';
 export * from './CreateFamily';
 export * from './CreateProject';
