+++ src/pages/index.ts
@@ -1,2 +1,2 @@
-export * from './Families';
-export * from './Projects';
+export * from './Dashboard/Families';
+export * from './Dashboard/Projects';
